issue 2721 refactor sls to implement restoring function WIP
